rename models package to model
