+++ src/main/java/com/springboot/task6/model/Product.java
