+++ src/main/java/com/springboot/task6/repositories/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.springboot.task6.repositories;
 
-import com.springboot.task6.models.Member;
+import com.springboot.task6.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
